Rename ingress module to ingress_client
